+++ InheritanceProject1/InheritanceProject1/Models/Tablet.cs
@@ -8,7 +8,7 @@ namespace InheritanceProject1.Models
 {
     public class Tablet : Computer
     {
-        public Tablet(string manufacturer, int cpuMhz, int cpuCores) : base(manufacturer, cpuMhz, cpuCores)
+        public Tablet(string manufacturer, float cpuMhz, int cpuCores) : base(manufacturer, cpuMhz, cpuCores)
         {
 
         }
